CFOPSolver.solve() method now can suppress the output messages, and just return the solution
ADDED: suppress_progress_messages kwarg to solve() to not print out any
of the updates about how the solve is going. It defaults to False, to
maintain expected behavior.
ADDED: error handling if the cube is already solved, it now pops up a
sugggestion that the cube may be solved and gracefully continues.